Move sudo use inside mkandroidsd script
The Android environment setup (envsetup.sh) applies only to the current
user.  As such, it is necessary to specify the path to call the
mkandroidsd script with sudo---it is no longer on the path. Instead,
we move the calls to sudo inside the script.  The user can call it
without sudo but will be prompted for sudo permissions.