Update uboot make procedure
Ensure u-boot is only re-made when the source files change and include it in
the default target.  As no other targets depend on the creation of MLO and
u-boot (these aren't in the tarballs), don't worry about including it as an
Android module---they can be built late---just as long as they're built.

Signed-off-by: Ash Charles <ashcharles@gmail.com>