Update linux make procedure (now out-of-tree)
Ensure linux is only re-made when the source files or configuration
file (now shipped in the device directory) change.  This allows for
zImage booting including modules, and firmware.

To be fixed...
We could add this to the default build target but, as the build output
needs to be installed in the system image, this would provide no
guarantee that the system image would get rebuilt when linux is
rebuilt.  For now:
 $ m linux
 $ m systemtarball

Signed-off-by: Ash Charles <ashcharles@gmail.com>